requirements.txt
streamlit>=1.10.0
openai>=1.0.0
pyyaml>=6.0
matplotlib>=3.5.0
pillow>=8.0.0
python-dotenv>=0.19.0
pdfkit>=1.0.0
httpx>=0.23.0
markdown>=3.4.0